Move randomization functions to bench_test.go.
