+++ kv_test.go
@@ -2,12 +2,9 @@ package kv_test
 
 import (
 	"bytes"
-	"encoding/binary"
 	"fmt"
 	"iter"
-	"math"
 	"math/bits"
-	rand "math/rand/v2"
 	"reflect"
 	"slices"
 	"strings"
@@ -242,45 +239,6 @@ func collect(itr iter.Seq2[[]byte, byte]) []entry {
 		entries = append(entries, entry{k, v})
 	}
 	return entries
-}
-
-func randomBytes(n int, random *rand.Rand) []byte {
-	if n == 0 {
-		return []byte{}
-	}
-	k := (n-1)/8 + 1
-	b := make([]byte, k*8)
-	for i := range k {
-		binary.BigEndian.PutUint64(b[i*8:], random.Uint64())
-	}
-	return b[:n]
-}
-
-func randomByte(random *rand.Rand) byte {
-	return byte(random.UintN(256))
-}
-
-// Returns a random key of with length chosen from a roughly normal distribution
-// with the given mean. Lengths will range from 0 to 2*mean.
-func randomKey(meanLen int, random *rand.Rand) []byte {
-	const bound = 4.0 // chosen experimentally
-	val := random.NormFloat64()
-	for val < -bound || val > +bound {
-		val = random.NormFloat64()
-	}
-	// val is in [-bound, +bound], translate that to [0, 2*mean]
-	val = (val + bound) * float64(meanLen) / bound
-	return randomBytes(int(math.Round(val)), random)
-}
-
-func randomFixedLengthKey(keyLen int, random *rand.Rand) []byte {
-	return randomBytes(keyLen, random)
-}
-
-func shuffle[S ~[]E, E any](slice S, random *rand.Rand) {
-	random.Shuffle(len(slice), func(i, j int) {
-		slice[i], slice[j] = slice[j], slice[i]
-	})
 }
 
 // storeConfigs for all possible subsequences of presentKeys.
